+++ variables.tf
@@ -3,10 +3,12 @@
 ###############
 
 variable "administrator_login" {
+  type        = string
   description = "The Administrator Login for the MySQL Flexible Server."
 }
 
 variable "administrator_password" {
+  type        = string
   description = "The Password associated with the administrator_login for the MySQL Flexible Server."
   sensitive   = true
 }
@@ -27,61 +29,79 @@ variable "firewall_rules" {
 }
 
 variable "geo_redundant_backup_enabled" {
-  description = "Is Geo-Redundant backup enabled on the MySQL Flexible Server."
   type        = bool
+  description = "Is Geo-Redundant backup enabled on the MySQL Flexible Server."
   default     = false
 }
 
 variable "location" {
+  type        = string
   description = "Specifies the supported Azure location where the resource exists."
   default     = "canadacentral"
 }
 
 variable "name" {
+  type        = string
   description = "The name of the MySQL Flexible Server."
 }
 
 variable "storage_account_name" {
+  type        = string
   description = "Name of the storage account used for diagnostics (optional, if not provided the name is auto-generated)."
   default     = null
 }
 
 variable "mysql_version" {
+  type        = string
   description = "The version of the MySQL Flexible Server."
   default     = "8.0.21"
 }
 
 variable "resource_group_name" {
+  type        = string
   description = "The name of the resource group in which to create the MySQL Flexible Server."
 }
 
 variable "sku_name" {
+  type        = string
   description = "Specifies the SKU Name for this MySQL Flexible Server."
   default     = "GP_Standard_D4ds_v4"
 }
 
 variable "storagesize_gb" {
-  description = "Specifies the version of MySQL to use."
+  type        = number
+  description = "Specifies the storage size in GB for the MySQL Flexible Server."
   default     = 128
+  validation {
+    condition     = var.storagesize_gb >= 20 && var.storagesize_gb <= 16384
+    error_message = "Valid values for storagesize_gb is between 20 and 16384."
+  }
 }
 
 variable "iops" {
+  type        = number
   description = "The storage IOPS for the MySQL Flexible Server."
+  validation {
+    condition     = var.iops >= 360 && var.iops <= 20000
+    error_message = "Valid values for iops is between 360 and 20000."
+  }
 }
 
 variable "tags" {
-  description = "A mapping of tags to assign to the resource."
   type        = map(string)
+  description = "A mapping of tags to assign to the resource."
   default = {
     environment : "dev"
   }
 }
 
 variable "project" {
+  type        = string
   description = "Name of client project"
 }
 
 variable "environment" {
+  type        = string
   description = "The environment used for keyvault access."
 }
 
@@ -90,15 +110,25 @@ variable "environment" {
 ##################
 
 variable "delegated_subnet_id" {
-  description = "The subnet where you want the database created. The subnet must be delegated to Microsoft.DBforMySQL/flexibleServers."
   type        = string
+  description = "The subnet where you want the database created. The subnet must be delegated to Microsoft.DBforMySQL/flexibleServers."
   default     = null
 }
 
 variable "private_dns_zone_id" {
-  description = "The ID of the private DNS zone to create the MySQL Flexible Server. The private DNS zone must end with the suffix .mysql.database.azure.com."
   type        = string
+  description = "The ID of the private DNS zone to create the MySQL Flexible Server. The private DNS zone must end with the suffix .mysql.database.azure.com."
   default     = null
+}
+
+variable "public_network_access" {
+  description = "(Optional) Specifies whether this MySQL Flexible Server is publicly accessible."
+  type        = string
+  default     = "Disabled"
+  validation {
+    condition     = contains(["Disabled", "Enabled"], var.public_network_access)
+    error_message = "Valid values for public_network_access are Disabled or Enabled."
+  }
 }
 
 variable "kv_private_endpoints" {
@@ -127,20 +157,21 @@ variable "kv_private_endpoints" {
   }
 }
 
-variable "public_network_access_enabled" {
-  description = "(Required) Whether or not public network access is allowed."
+variable "kv_public_network_access_enabled" {
+  type        = bool
+  description = "(Required) Whether or not public network access is allowed for the key vault."
   default     = false
 }
 
 variable "kv_subnet_ids" {
-  description = "The subnets for the key vault."
   type        = list(string)
+  description = "The subnets for the key vault."
   default     = null
 }
 
 variable "sa_subnet_ids" {
-  description = "The subnets for the storage account."
   type        = list(string)
+  description = "The subnets for the storage account."
   default     = null
 }
 
@@ -153,15 +184,13 @@ variable "diagnostics" {
   type = object({
     destination   = string
     eventhub_name = string
-    logs          = list(string)
-    metrics       = list(string)
   })
   default = null
 }
 
 variable "sa_create_log" {
-  description = "Creates a storage account to be used for diagnostics logging of the MySQL database created if the variable is set to `true`."
   type        = bool
+  description = "Creates a storage account to be used for diagnostics logging of the MySQL database created if the variable is set to `true`."
   default     = false
 }
 
@@ -176,23 +205,28 @@ variable "sa_create_log" {
 ######################################################################
 
 variable "kv_pointer_enable" {
+  type        = bool
   description = "Flag kv_pointer_enable can either be `true` (state from key vault), or `false` (state from terraform)."
   default     = false
 }
 
 variable "kv_pointer_name" {
+  type        = string
   description = "The key vault name to be used when kv_pointer_enable is set to `true`."
   default     = null
 }
 
 variable "kv_pointer_rg" {
+  type        = string
   description = "The key vault resource group to be used when kv_pointer_enable is set to `true`."
   default     = null
 }
 
 variable "kv_pointer_sqladmin_password" {
+  type        = string
   description = "The sqladmin password to be looked up in key vault when kv_pointer_enable is set to `true`."
   default     = null
+  sensitive   = true
 }
 
 ##################
@@ -203,7 +237,7 @@ variable "mysql_configurations" {
   type = map(string)
   default = {
     audit_log_enabled       = "ON"
-    audit_log_events        = "CONNECTION_V2, ADMIN"
+    audit_log_events        = "CONNECTION_V2, ADMIN, DDL, DCL, DML_NONSELECT"
     max_connect_errors      = "20"
     innodb_buffer_pool_size = "12884901888"
     max_allowed_packet      = "536870912"